# ogr2ogr: speed-up reprojection in Arrow code path