… autotest/cpp/test_ogr.cpp
… autotest/cpp/test_ogr_wkb.cpp
… autotest/utilities/test_ogr2ogr_lib.py
… ogr/ogr_wkb.cpp
… ogr/ogrct.cpp
… ogr/ogrgeometryfactory.cpp
… ogr/ogrsf_frmts/gpkg/ogr_geopackage.h
… ogr/ogrsf_frmts/gpkg/ogrgeopackagedatasource.cpp
… ogr/ogrsf_frmts/gpkg/ogrgeopackageutility.cpp
… ogr/ogrsf_frmts/gpkg/ogrgeopackageutility.h