(on top of PR #10716)

Use newly added OGRWKBTransform(), when possible, to speed-up -t_srs in Arrow code path (which is triggered in Parquet/GPKG -> anything, when no option or no other option than -t_srs is specified). Also use multi-threaded coordinate transformation by splitting the features within each batch in several sub-batches, each processed in its own thread.

Can result in a 3.3x speed-up in ogr2ogr Parquet->Parquet -t_srs use case on a 3.2 million feature dataset:

No reprojection:
```
$ time ogr2ogr out.parquet nz-building-outlines.parquet
real	0m5,687s
user	0m5,254s
sys	0m1,032s
```

Reprojection EPSG:2193 ("NZGD2000 / New Zealand Transverse Mercator 2000") to EPSG:4326 without optimization:
```
$ time ogr2ogr out.parquet nz-building-outlines.parquet -t_srs EPSG:4326 --config OGR2OGR_USE_ARROW_API=NO
real	0m24,038s
user	0m23,472s
sys	0m1,314s
```

Reprojection *with* optimization using 6 theads for reprojection:
```
$ time ogr2ogr out.parquet nz-building-outlines.parquet -t_srs EPSG:4326
real	0m7,100s
user	0m14,376s
sys	0m1,044s
```

Reprojection *with* optimization using 1 thead for reprojection:
```
$ time ogr2ogr out.parquet nz-building-outlines.parquet -t_srs EPSG:4326 --config GDAL_NUM_THREADS=1
real	0m13,773s
user	0m12,786s
sys	0m1,002s
```

Also applies for GPKG -> GPKG  with a 2x speedup:
```
$ time ogr2ogr out.gpkg nz-building-outlines.gpkg -t_srs EPSG:4326
real	0m17,976s
user	0m29,195s
sys	0m1,802s
```

Without optimization:
```
$ time ogr2ogr out.gpkg nz-building-outlines.gpkg -t_srs EPSG:4326 --config OGR2OGR_USE_ARROW_API=NO
real	0m36,688s
user	0m35,809s
sys	0m2,269s
```